minus supermassive-ast
