+++ packages/supermassive/src/executeWithSchema.ts
@@ -2,7 +2,7 @@ import { buildASTSchema } from "graphql";
 import { executeWithoutSchema } from "./index";
 import { PromiseOrValue } from "./jsutils/PromiseOrValue";
 import { ExecutionResult, ExecutionWithSchemaArgs } from "./types";
-import { extractMinimalViableSchemaForRequestDocument } from "./supermassive-ast/addMinimalViableSchemaToRequestDocument";
+import { extractMinimalViableSchemaForRequestDocument } from "./utilities/addMinimalViableSchemaToRequestDocument";
 
 export function executeWithSchema({
   typeDefs,
