--- packages/supermassive/src/supermassive-ast/index.ts
@@ -1,61 +1,0 @@
-export type {
-  NameNode,
-  DocumentNode,
-  OperationDefinitionNode,
-  VariableDefinitionNode,
-  VariableNode,
-  SelectionSetNode,
-  FieldNode,
-  ArgumentNode,
-  FragmentSpreadNode,
-  InlineFragmentNode,
-  FragmentDefinitionNode,
-  IntValueNode,
-  FloatValueNode,
-  StringValueNode,
-  BooleanValueNode,
-  NullValueNode,
-  EnumValueNode,
-  ListValueNode,
-  ObjectValueNode,
-  ObjectFieldNode,
-  DirectiveNode,
-  NamedTypeNode,
-  ListTypeNode,
-  NonNullTypeNode,
-  SchemaDefinitionNode,
-  OperationTypeDefinitionNode,
-  ScalarTypeDefinitionNode,
-  ObjectTypeDefinitionNode,
-  FieldDefinitionNode,
-  InputValueDefinitionNode,
-  InterfaceTypeDefinitionNode,
-  UnionTypeDefinitionNode,
-  EnumTypeDefinitionNode,
-  EnumValueDefinitionNode,
-  InputObjectTypeDefinitionNode,
-  DirectiveDefinitionNode,
-  SchemaExtensionNode,
-  ScalarTypeExtensionNode,
-  ObjectTypeExtensionNode,
-  InterfaceTypeExtensionNode,
-  UnionTypeExtensionNode,
-  EnumTypeExtensionNode,
-  InputObjectTypeExtensionNode,
-  SelectionNode,
-  TypeNode,
-  OperationTypeNode,
-} from "graphql";
-// export { addMinimalViableSchemaToRequestDocument } from "./addMinimalViableSchemaToRequestDocument";
-export { annotateDocumentGraphQLTransform } from "./annotateDocumentGraphQLTransform";
-
-export {
-  specifiedDirectives,
-  isSpecifiedDirective,
-  GraphQLIncludeDirective,
-  GraphQLSkipDirective,
-  GraphQLDeprecatedDirective,
-  GraphQLSpecifiedByDirective,
-  GraphQLDeferDirective,
-  GraphQLStreamDirective,
-} from "../types/directives";
